code refactoring, hard coded game data added with game component
