+++ src/app/app.component.html
@@ -1,4 +1,2 @@
 <app-menu></app-menu>
-<app-header></app-header>
 <router-outlet></router-outlet>
-<app-footer></app-footer>
