+++ src/app/pages/admin/admin.module.ts
@@ -8,11 +8,12 @@ import { ActionsComponent } from './actions/actions.component';
 import { ChartsComponent } from './charts/charts.component';
 import { CodeqlComponent } from './codeql/codeql.component';
 import { TimeComponent } from './time/time.component';
+import { GamesComponent } from './games/games.component';
 
 const routes: Routes = [{ path: '', component: AdminComponent }];
 
 @NgModule({
-  declarations: [AdminComponent, ListComponent, ActionsComponent, ChartsComponent, CodeqlComponent, TimeComponent],
+  declarations: [AdminComponent, ListComponent, ActionsComponent, ChartsComponent, CodeqlComponent, TimeComponent, GamesComponent],
   imports: [CommonModule, MaterialModule, RouterModule.forChild(routes)],
   exports: [RouterModule],
 })
