+++ src/app/pages/admin/games/games.component.cy.ts
@@ -1,0 +1,7 @@
+import { GamesComponent } from './games.component'
+
+describe('GamesComponent', () => {
+  it('should mount', () => {
+    cy.mount(GamesComponent)
+  })
+})
