+++ src/app/pages/admin/games/games.component.html
@@ -1,0 +1,1 @@
+<p>games works!</p>
