--- src/app/skeleton/footer/footer.component.sass
@@ -1,73 +1,0 @@
-.resume
-    position: relative
-    overflow: hidden
-    width: 100%
-    padding-top: 56.25%
-
-    &__iframe
-        position: absolute
-        top: 0
-        left: 0
-        bottom: 0
-        right: 0
-        width: 100%
-        height: 100%
-
-.card-container .material-icons path
-    fill: rgb(105, 103, 103)
-
-.card-container
-    display: flex
-    justify-content: center
-    align-items: center
-    flex-direction: column
-
-    a
-        color: #fff
-
-    & .card:not(:last-child)
-        margin-right: 0
-
-.card-container .card:not(.highlight-card)
-    cursor: pointer
-
-.card-container .card:not(.highlight-card):hover
-    transform: translateY(-3px)
-    box-shadow: 0 4px 17px rgba(0, 0, 0, 0.35)
-
-.card-container .card:not(.highlight-card):hover .material-icons path
-    fill: rgb(105, 103, 103)
-
-.card-container
-    &__link-container
-        display: flex
-        margin: 0.25rem
-        align-items: center
-    &__policy-container
-        display: flex
-        color: #fff
-        align-items: center
-
-footer
-    position: relative
-    height: fit-content
-    background: #000
-    margin: 1vh 0 0 0
-
-    a
-        margin: .1rem
-        height: 2rem
-        width: 2rem
-        border-radius: 50%
-        background-color: white
-        border: 1px solid #eeeeee
-        display: flex
-        justify-content: center
-        align-items: center
-        cursor: pointer
-        box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24)
-        transition: 1s ease-out
-
-        &:hover
-            transform: translateY(-0.25rem)
-            box-shadow: 0px 3px 15px rgba(0, 0, 0, 0.2)
