--- src/app/skeleton/header/header.component.sass
@@ -1,43 +1,0 @@
-.container
-    min-width: 100vw
-    background-size: 150%
-    background-position: center
-    background-repeat: no-repeat
-    background: #ffd740
-    &__name
-        font-size: 3rem
-        justify-content: center
-        color: #fff
-        // color: #e1d3d3bd
-    &__details
-        justify-content: center
-        font-size: 1.5rem
-        margin-top: -.8rem
-        color: #fff
-        // color: #e1d3d3bd
-
-    @media screen and (max-width: 535px)
-        .container
-            &__name
-                font-size: 1.35rem
-                margin-top: -0.25rem
-            &__details
-                font-size: .75rem
-                margin: -1.5rem 0px -1rem 0
-
-    @media screen and (min-width: 1330px)
-        .container
-            &__name
-                font-size: 3.75rem
-
-    @media screen and (min-width: 2400px) and (max-width: 3000px)
-        .container
-            &__details
-                font-size: 1.8rem
-
-    @media screen and (min-width: 3000px)
-        .container
-            &__name
-                font-size: 4.5rem
-            &__details
-                font-size: 2rem
