+++ src/assets/models/models.interface.ts
@@ -2,7 +2,6 @@ export interface Common {
   title: string;
   url: string;
 }
-
 export interface Biodata {
   name: string;
   role: string;
@@ -13,14 +12,12 @@ export interface Biodata {
   medium: string;
   hashnode: string;
 }
-
 export interface SCMData {
   starredRepoCount: number;
   openIssueCount: number;
   list: Common[];
   scmActivity: SCMActivity[];
 }
-
 export interface SCMActivity {
   loc: number;
   pr: number;
@@ -31,7 +28,13 @@ export interface WeekData {
   Key: string;
   Value: string;
 }
-
 export interface Token {
   token: string;
+}
+
+export interface Game{
+  gamedetails: string;
+  pgn: string;
+  result: string;
+  gameurl: string;
 }
